+++ src/base/PluginTrait.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * @link      https://dukt.net/facebook/
- * @copyright Copyright (c) 2019, Dukt
+ * @copyright Copyright (c) 2021, Dukt
  * @license   https://github.com/dukt/facebook/blob/master/LICENSE.md
  */
 
@@ -23,6 +23,17 @@ use dukt\facebook\Plugin as Facebook;
  */
 trait PluginTrait
 {
+    /**
+     * Returns the accounts service.
+     *
+     * @return \dukt\facebook\services\Accounts The accounts service
+     */
+    public function getAccounts()
+    {
+        /** @var Facebook $this */
+        return $this->get('accounts');
+    }
+
     /**
      * Returns the api service.
      *
@@ -90,36 +101,37 @@ trait PluginTrait
         return false;
     }
 
-    public function getClientId()
+    /**
+     * Gets the OAuth client ID
+     *
+     * @param bool $parse
+     * @return bool|mixed|string|null
+     */
+    public function getClientId(bool $parse = true)
     {
         $clientId = Facebook::$plugin->getSettings()->oauthClientId;
 
-        if($clientId) {
-            return $clientId;
-        } else {
-            $plugin = Craft::$app->getPlugins()->getPlugin('facebook');
-            $settings = $plugin->getSettings();
-
-            if(!empty($settings['oauthClientId']))
-            {
-                return $settings['oauthClientId'];
-            }
+        if (!$clientId) {
+            return null;
         }
+
+        return $parse ? Craft::parseEnv($clientId) : $clientId;
     }
-    public function getClientSecret()
+
+    /**
+     * Gets the OAuth client secret
+     *
+     * @param bool $parse
+     * @return bool|mixed|string|null
+     */
+    public function getClientSecret(bool $parse = true)
     {
         $clientSecret = Facebook::$plugin->getSettings()->oauthClientSecret;
 
-        if($clientSecret) {
-            return $clientSecret;
-        } else {
-            $plugin = Craft::$app->getPlugins()->getPlugin('facebook');
-            $settings = $plugin->getSettings();
-
-            if(!empty($settings['oauthClientSecret']))
-            {
-                return $settings['oauthClientSecret'];
-            }
+        if (!$clientSecret) {
+            return null;
         }
+
+        return $parse ? Craft::parseEnv($clientSecret) : $clientSecret;
     }
 }