+++ src/controllers/SettingsController.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * @link      https://dukt.net/facebook/
- * @copyright Copyright (c) 2019, Dukt
+ * @copyright Copyright (c) 2021, Dukt
  * @license   https://github.com/dukt/facebook/blob/master/LICENSE.md
  */
 
@@ -74,7 +74,8 @@ class SettingsController extends Controller
 
         return $this->renderTemplate('facebook/settings/oauth', [
             'redirectUri' => Facebook::$plugin->oauth->getRedirectUri(),
-            'settings' => $plugin->getSettings(),
+            'oauthClientId' => Facebook::$plugin->getClientId(false),
+            'oauthClientSecret' => Facebook::$plugin->getClientSecret(false),
         ]);
     }