+++ src/errors/InvalidAccountException.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * @link      https://dukt.net/facebook/
+ * @copyright Copyright (c) 2021, Dukt
+ * @license   https://github.com/dukt/facebook/blob/master/LICENSE.md
+ */
+
+namespace dukt\facebook\errors;
+
+use yii\base\Exception;
+
+/**
+ * Class InvalidAccountException
+ */
+class InvalidAccountException extends Exception
+{
+}